chore: Implement video stream priority for native videos (#62)
chore: Implement video stream priority for native videos

- Handle video streaming priority based on user preference

fix: LEARNER-10260

+++ core/src/main/java/org/openedx/core/module/DownloadWorker.kt
@@ -147,7 +147,7 @@ class DownloadWorker(
                     DownloadModelEntity.createFrom(
                         downloadTask.copy(
                             downloadedState = DownloadedState.DOWNLOADED,
-                            size = File(downloadTask.path).length().toInt(),
+                            size = File(downloadTask.path).length(),
                             transcriptDownloadedStatus = if (isTranscriptsDownloaded) {
                                 TranscriptsDownloadedState.DOWNLOADED
                             } else {
